requirements/base.txt
Django==4.2
django-bootstrap4
django-bootstrap-datepicker-plus
setuptools
